Fixed integer rendering

+++ DCode.Services/Common/ICommonService.cs
@@ -47,7 +47,7 @@ namespace DCode.Services.Common
 
         void MigrateGamificationRecords();
 
-        decimal? GetApprovedApplicantHours();
+        int? GetApprovedApplicantHours();
 
         int? GetUserPoints();
